test: add JuliaSimCompiler only to run its integration tests
Remove `JuliaSimCompiler` as a test dependency as that would require
having access to the package at all times to run the tests, making it
synonymous with a direct dependency (and less of an extension) in the
context of running tests.
Instead, explicitly `Pkg.add` `JuliaSimCompiler` instead in the
relevant test matrices.